Implementação Aula 03/10/23

+++ backend/src/controllers/topic.controller.ts
@@ -1,8 +1,8 @@
-import { Body, Controller, Delete, Get, HttpCode, Param, ParseIntPipe, Post } from "@nestjs/common";
-import { Topic } from "src/entities/topic.entity";
+import { Topic } from './../entities/topic.entity';
+import { Body, Controller, Delete, Get, HttpCode, HttpException, HttpStatus, Param, ParseIntPipe, Post, Put } from "@nestjs/common";
 import { TopicService } from "src/services/topic.service";
 
-@Controller('topics')
+@Controller('Topics')
 export class TopicController {
 
     constructor(private readonly service: TopicService){}
@@ -14,18 +14,45 @@ export class TopicController {
 
     @Get(':id')
     findById(@Param('id', ParseIntPipe) id: number): Promise<Topic> {
-        return this.service.findById(id)
+        const found = this.service.findById(id);
+
+        if(!found){
+            throw new HttpException('Topic not found', HttpStatus.NOT_FOUND)
+        }
+
+        return found;
     }
 
+
     @Post()
-    create(@Body() topic: Topic): Promise<Topic> {
-        return this.service.create(topic)
+    create(@Body() Topic: Topic): Promise<Topic> {
+        return this.service.create(Topic)
     }
 
     @Delete(':id')
     @HttpCode(204)
     delete(@Param('id', ParseIntPipe) id: number): Promise<void>  {
+        const found = this.service.findById(id);
+
+        if(!found){
+            throw new HttpException('Topic not found', HttpStatus.NOT_FOUND)
+        }
+
         return this.service.delete(id);
+
     }
+
+    @Put(':id')
+    async update(@Param('id', ParseIntPipe) id: number, @Body() Topic: Topic): Promise<Topic> {
+        const found = this.service.findById(id);
+
+        if(!found){
+            throw new HttpException('Topic not found', HttpStatus.NOT_FOUND)
+        }
+
+        return this.service.update(id,Topic);
+    }
+
+
 
 }
